app/requirements.txt
Flask==3.1.2
cryptography==45.0.7
alembic==1.16.5
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask-Admin==1.6.1
flask-basicauth==0.2.0
Flask-Cors==6.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
Pillow==11.3.0
PyMySQL==1.1.2
SQLAlchemy==2.0.43
typing-extensions==4.15.0
Werkzeug==3.1.3
WTForms==2.3.0
python-dotenv==1.1.1
gunicorn